backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-core
langchain-community
pypdf
faiss-cpu
tiktoken
python-dotenv
requests  # Added for API calls to Grok